event CRUD finished
event CRUD finished, join event need debug.

+++ backend/demoApp/src/main/java/com/natalie/demoApp/repos/MemberRepository.java
@@ -3,6 +3,6 @@ package com.natalie.demoApp.repos;
 import com.natalie.demoApp.entity.Member;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MemberRepository extends CrudRepository<Member,Integer>{
+public interface MemberRepository extends CrudRepository<Member,Long>{
 
 }
